Respect requested locale while creating codes
The CodeTypeCacheKey which is used to request an ICodeType contains a
Locale that needs to be respected while resolving the ICodeType. If e.g.
the current locale is ENGLISH but the ICodeType was requested for GERMAN
the ICodeType needs to be inside a GERMAN RunContext to ensure correct
translations. Otherwise, there will be a GERMAN ICodeType in the cache
for an ENGLISH CodeTypeCacheKey resulting in e.g. wrong translations for
all english users.